+++ src/main/java/de/is24/deadcode4j/plugin/FindDeadCodeOnlyMojo.java
@@ -132,6 +132,7 @@ public class FindDeadCodeOnlyMojo extends AbstractSlf4jMojo {
     private DeadCode analyzeCode() throws MojoExecutionException {
         Set<Analyzer> analyzers = Sets.<Analyzer>newHashSet(
                 new AopXmlAnalyzer(),
+                new ApacheTilesAnalyzer(),
                 new CastorClassesAnalyzer(),
                 new ClassDependencyAnalyzer(),
                 new FacesConfigXmlAnalyzer(),